First conversion to sdl3

+++ Minigin/Font.cpp
@@ -1,12 +1,12 @@
 #include <stdexcept>
-#include <SDL_ttf.h>
+#include <SDL3_ttf/SDL_ttf.h>
 #include "Font.h"
 
 TTF_Font* dae::Font::GetFont() const {
 	return m_font;
 }
 
-dae::Font::Font(const std::string& fullPath, unsigned int size) : m_font(nullptr)
+dae::Font::Font(const std::string& fullPath, float size) : m_font(nullptr)
 {
 	m_font = TTF_OpenFont(fullPath.c_str(), size);
 	if (m_font == nullptr)
